feat(packages): add detailed package view with itinerary and inclusions
- Add show method in LandingController to load and transform a single package details
- Implement package inclusions logic based on accommodations and type
- Build itinerary display using package itineraries or default template
- Create reusable React components for package detail hero, content, itinerary, and inclusions
- Update packages list and showcase components to use numeric price and format with currency code
- Add booking card component with quantity selector and pricing details
- Enhance landing hero and navigation with links and theme toggle
- Fix total revenue type casting in admin and staff dashboards
- Update admin dashboard with better pie chart labeling and currency formatting
- Add new dashboard page with charts for admin summary
- Change packages page title to simpler "Packages"

+++ app/Http/Controllers/Admin/DashboardController.php
@@ -30,7 +30,7 @@ class DashboardController extends Controller implements HasMiddleware
         $stats = [
             'total_packages' => Package::count(),
             'total_bookings' => Booking::count(),
-            'total_revenue' => Transaction::where('status', 'success')->sum('amount'),
+            'total_revenue' => (float) Transaction::where('status', 'success')->sum('amount'),
             'total_users' => User::count(),
         ];
